Mostly player fixes.
Removed keyboard turning controls from the player.
Removed redundant code from PlayerController.
Normalized the rotation of the player to be in line with Unity's coordinate system.
Adjusted player controller script to match player model rotation normalization.
Moved movment code to Update() to reduce stuttering; this is subject to change if physics and angular momentum based movement is introduced later.
Removed the stamina restoring capability of score turn-ins.
Removed the ability of the player to have a permanent speed boost when holding down the 'space' key.
Added an extra variable to the player that is the delay in seconds before stamina begins to recover.
Added a new variable to the player that is the minimum stamina required to initiate a new speed boost.
Added extra public variables concerning the player stats to the PlayerController script, and added matching functions to adjust the GameController's information about the player (maxStamina, maxHealth, etc).